+++ packages/kit/test/apps/basics/src/routes/store/stuff/__error.svelte
@@ -1,5 +1,5 @@
 <script context="module">
-	/** @type {import('@sveltejs/kit').ErrorLoad} */
+	/** @type {import('@sveltejs/kit').Load} */
 	export function load({ error }) {
 		return {
 			stuff: {
@@ -9,3 +9,10 @@
 		};
 	}
 </script>
+
+<script>
+	import { page } from '$app/stores';
+	import { goto } from '$app/navigation';
+</script>
+
+<button id="reload-button" on:click={() => goto($page.url.toString())}>Reload</button>